backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain-openai==0.3.0
langchain-core>=0.3.15,<0.4.0
python-dotenv==1.0.1
pillow==10.4.0
python-multipart==0.0.9
